Add files via upload
Description: This C program implements Selection Sort, which is a sorting algorithm that sorts an array by repeatedly finding the minimum element from the unsorted part of the array and putting it at the beginning. It has a time complexity of O(n^2).

Algorithm:

Initialize min_idx as the first element of the array. Traverse the array from the second element to the last. For each element, find the minimum element from the unsorted part of the array and swap it with the current element. Repeat the process until the array is sorted. Input: The input to this program is an array of integers.

Output: The output of this program is the sorted array.